create metadata on serverside page.tsx
